Use whitelist to restrict permission to user profile override
Using the tenant service with a feature that MUST be named
`user.update-tenant` and which should be configured with
the user ids, in order to restrict permission to use the content
of `contextInformation.tenantConfig` in the user's profile in `auth`
to change the default versions of Che/Jenkins/etc.

Include a `Toggle Client` that can use a mock unleashclient for testing,
as well as an http client in the controller for mocking the interactions
with `auth`.

Signed-off-by: Xavier Coulon <xcoulon@redhat.com>