API to delete/deprovision a user from Tenant (OSIO#3338)
Refactored the `tenants` controller to include a new `openshift`
service that provides a method to delete a namespace using the
OSO API. This service can use a different HTTP client transport,
which allows for testing with the `go-vcr` library.

If a namespace deletion fails, the whole deletion stops, except for
`403 Forbidden` responses from OSO which can be returned if the
namespace to delete does not exist. In that case, there's no reason
to halt the tenant deletion.

Fixes openshiftio/openshift.io#3338

Signed-off-by: Xavier Coulon <xcoulon@redhat.com>