chore: refactoring of the components that comunicate with auth service
This is the first step of a complete rewrite of the tenant service and
is a preparation for OSIO on OSD support.
It "only" moves the code and integrates together those parts that are
somehow related to auth service. Everything that is related to getting information about users
and token management is moved to auth package and the main entry point
for the communication is auth-service.
Everything that is related to clusters live in cluster package and the main entry point is
cluster-service.
Previously, there was used a combination of resolve functions that were
integrated in each other and internaly used. Instead of it, I'm proposing
using a service objects and passing the objects to the methods.
As part of this refactoring I was trying to use the fabric8-common repository when possible.
Tests are techincally almost same as in previous version, there was just
slightly changed the setup of the tests.