chore: refactoring of the components that comunicate with auth service (#632)
This is the first step of a complete rewrite of the tenant service and is a preparation for OSIO on OSD support.
It moves the code and integrates together those parts that are somehow related to auth service. Everything that is related to getting information about users and token management is moved to auth package and the main entry point for the communication is auth-service.
Everything that is related to clusters lives in the cluster package and the main entry point is cluster service.