feat: simplify template management (#634)
The location of the templates is changed from the external projects f8-tenant-[che|jenkins|team] to the directory environment/templates/ inside of this fabric8-tenant project. The files (templates) located in the directory and their names correspond to namespaces they are used for. All objects (contained in the templates) have been moved from the original ones (downloaded from Maven central) just with small cosmetic changes - the result and the number of the objects are same.